huazie_flea-frame-cache部分代码重构，新增Redis客户端策略_MOD
