+++ flea-frame-cache/src/main/java/com/huazie/frame/cache/core/manager/CoreSpringCacheManager.java
@@ -1,4 +1,4 @@
-package com.huazie.frame.cache.core;
+package com.huazie.frame.cache.core.manager;
 
 import com.huazie.frame.cache.AbstractSpringCache;
 import com.huazie.frame.cache.AbstractSpringCacheManager;
