+++ flea-frame-config/src/main/resources/applicationContext.xml
@@ -108,7 +108,7 @@
         配置缓存管理MemCachedSpringCacheManager
         配置缓存时间 configMap (key缓存对象名称 value缓存过期时间)
     -->
-    <bean id="memCachedSpringCacheManager" class="com.huazie.frame.cache.memcached.MemCachedSpringCacheManager">
+    <bean id="memCachedSpringCacheManager" class="com.huazie.frame.cache.memcached.manager.MemCachedSpringCacheManager">
         <property name="configMap">
             <map>
                 <entry key="fleaconfigdata" value="86400"/>
@@ -124,7 +124,7 @@
         配置缓存管理 redisShardedSpringCacheManager
         配置缓存时间 configMap (key缓存对象名称 value缓存过期时间)
     -->
-    <bean id="redisShardedSpringCacheManager" class="com.huazie.frame.cache.redis.RedisShardedSpringCacheManager">
+    <bean id="redisShardedSpringCacheManager" class="com.huazie.frame.cache.redis.manager.RedisShardedSpringCacheManager">
         <property name="configMap">
             <map>
                 <entry key="fleaconfigdata" value="86400"/>
@@ -140,7 +140,7 @@
         配置缓存管理 redisClusterSpringCacheManager
         配置缓存时间 configMap (key缓存对象名称 value缓存过期时间)
     -->
-    <bean id="redisClusterSpringCacheManager" class="com.huazie.frame.cache.redis.RedisClusterSpringCacheManager">
+    <bean id="redisClusterSpringCacheManager" class="com.huazie.frame.cache.redis.manager.RedisClusterSpringCacheManager">
         <property name="configMap">
             <map>
                 <entry key="fleamenufavorites" value="100"/>
@@ -149,7 +149,7 @@
     </bean>
 
     <!-- 配置核心Flea缓存管理类 RedisSpringCacheManager -->
-    <bean id="coreSpringCacheManager" class="com.huazie.frame.cache.core.CoreSpringCacheManager" />
+    <bean id="coreSpringCacheManager" class="com.huazie.frame.cache.core.manager.CoreSpringCacheManager" />
 
     <!-- 开启缓存 -->
     <cache:annotation-driven cache-manager="coreSpringCacheManager" proxy-target-class="true"/>
