+++ flea-frame-core/src/test/java/com/huazie/frame/core/base/cfgdata/FleaConfigDataSVImplTest.java
@@ -6,8 +6,8 @@ import com.huazie.frame.cache.AbstractSpringCache;
 import com.huazie.frame.cache.AbstractSpringCacheManager;
 import com.huazie.frame.cache.common.CacheEnum;
 import com.huazie.frame.cache.common.FleaCacheManagerFactory;
-import com.huazie.frame.cache.memcached.MemCachedSpringCacheManager;
-import com.huazie.frame.cache.redis.RedisShardedSpringCacheManager;
+import com.huazie.frame.cache.memcached.manager.MemCachedSpringCacheManager;
+import com.huazie.frame.cache.redis.manager.RedisShardedSpringCacheManager;
 import com.huazie.frame.common.EntityStateEnum;
 import com.huazie.frame.common.FleaFrameManager;
 import com.huazie.frame.common.slf4j.FleaLogger;
